# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: api-digitalmailsender
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8740
paths:
  /sendDigitalMail:
    post:
      tags:
      - Digital Mail
      operationId: sendDigitalMail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DigitalMailRequest'
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigitalMailResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api-docs:
    get:
      tags:
      - API
      summary: OpenAPI
      operationId: getApiDocs
      responses:
        "200":
          description: OK
          content:
            application/yaml:
              schema:
                type: string
      x-auth-type: None
      x-throttling-tier: Unlimited
      x-wso2-mutual-ssl: Optional
components:
  schemas:
    Attachment:
      required:
      - body
      - contentType
      - filename
      type: object
      properties:
        contentType:
          type: string
          description: "Allowed type is: application/pdf"
          example: application/pdf
        body:
          type: string
          description: Base64-encoded body
        filename:
          type: string
          description: The name of the file
          example: sample.pdf
      description: "An attachment, e.g. pdf, docx etc."
    BodyInformation:
      required:
      - body
      - contentType
      type: object
      properties:
        contentType:
          type: string
          description: "The content type for the message, text/plain for only text,\
            \ text/html for html messages."
          example: text/html
        body:
          type: string
          description: "Plain text if contentType is set to 'text/plain', BASE64-encoded\
            \ if contentType is set to 'text/html."
          example: PCFET0NUWVBFIGh0bWw+PGh0bWwgbGFuZz0iZW4iPjxoZWFkPjxtZXRhIGNoYXJzZXQ9InV0Zi04Ij48dGl0bGU+VGVzdDwvdGl0bGU+PC9oZWFkPjxib2R5PjxwPkhlbGxvPC9wPjwvYm9keT48L2h0bWw+
      description: The body of the message
    DigitalMailRequest:
      required:
      - headerSubject
      - municipalityId
      - partyId
      - supportInfo
      type: object
      properties:
        partyId:
          type: string
          description: partyId for the person or organization the digital mail should
            be sent to
          example: 6a5c3d04-412d-11ec-973a-0242ac130003
        municipalityId:
          type: string
          description: MunicipalityId
          example: "2281"
        headerSubject:
          type: string
          description: The subject of the digital mail.
          example: Viktig information från Sundsvalls kommun
        supportInfo:
          $ref: '#/components/schemas/SupportInfo'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        bodyInformation:
          $ref: '#/components/schemas/BodyInformation'
    SupportInfo:
      required:
      - contactInformationEmail
      - contactInformationPhoneNumber
      - contactInformationUrl
      - supportText
      type: object
      properties:
        supportText:
          type: string
          description: Information text describing the different ways the recipient
            may contact the sender.
          example: Kontakta oss via epost eller telefon.
        contactInformationUrl:
          type: string
          description: Url where the recipient may find more information.
          example: https://sundsvall.se/
        contactInformationPhoneNumber:
          type: string
          description: Phone number the recipient may call to get in contact with
            the sender.
          example: "4660191000"
        contactInformationEmail:
          type: string
          description: Email address the recipient may use to get in contact with
            the sender.
          example: sundsvalls.kommun@sundsvall.se
      description: Contains contact information and where the recipient may turn to
        for questions.
    DeliveryStatus:
      type: object
      properties:
        delivered:
          type: boolean
        partyId:
          type: string
        transactionId:
          type: string
    DigitalMailResponse:
      type: object
      properties:
        deliveryStatus:
          $ref: '#/components/schemas/DeliveryStatus'
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: '#/components/schemas/StatusType'
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
  securitySchemes: {}